split chunker file
